[refactor] mono -> restaurant-exposure
