+++ service/restaurant-exposure-service/src/main/java/woowa/team4/bff/RestaurantExposureServiceApplication.java
@@ -1,4 +1,4 @@
-package woowa.team4.bff.restaurantexposureservice;
+package woowa.team4.bff;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
