+++ service/restaurant-exposure-service/src/main/java/woowa/team4/bff/restauarntexposureservice/exposure/controller/get/GetRestaurantSummaryRequest.java
@@ -1,4 +1,4 @@
-package woowa.team4.bff.exposure.controller.get;
+package woowa.team4.bff.restauarntexposureservice.exposure.controller.get;
 
 public record GetRestaurantSummaryRequest(String keyword, String deliveryLocation, Integer pageNumber) {
 }
