+++ service/restaurant-exposure-service/src/main/resources/application.yml
@@ -4,8 +4,6 @@ server:
 spring:
   profiles:
     active: local
-  application:
-    name: "restaurant-service"
 
 ---
 spring:
